backend/requirements/base.txt
# Core Dependencies
fastapi>=0.95.0,<1.0.0
uvicorn[standard]>=0.21.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=1.0.0,<2.0.0
email-validator>=2.0.0,<3.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.10.0,<2.0.0
psycopg2-binary>=2.9.5,<3.0.0

# Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0

# Utilities
python-dateutil>=2.8.2,<3.0.0
pytz>=2022.7.1,<2024.0.0
python-slugify>=7.0.0,<8.0.0
pydantic[email]>=1.10.7,<2.0.0

# Logging
python-json-logger>=2.0.6,<3.0.0
structlog>=22.3.0,<24.0.0

# Async
anyio>=3.6.2,<4.0.0
